refactor: moved `ModelOperations` to `uql.models`
